requirements.txt
pytorch
pyro-ppl
matplotlib
numpy